pre-review method dist update
